Dia 0.97.1
Update diagrams README mentioning that the correct version of Dia to be
used is `0.97.1`. Tests under Windows revealed that Dia `>=0.97.2` will
draw zero-width borders when exporting to Cairo SVG (see #12).

Possibly, the problem is related to the Cairo library version employed,
not to Dia itself.

We're still unsure as to whether this Dia version problem also affects
builds for other OSs, but since I'm building the project under Windows
we're adding the Dia `0.97.1` restriction. More tests needed.