Dia Downloader for Windows
Add to diagrams' source folder a batch script that will download the
correct version of Dia for Windows from SourceForge and unpack it to
`assets_src/img/dia/Dia/`. (see #12)

Edit `.gitignore` file to ignore the downloaded Zip file and the `Dia/`
subfolder to which it's unpacked.

Tweak the diagrams `build.sh` script so that, if running under Windows,
it will check for the presence of the unpacked Dia, and use it instead.
The script now also checks that Dia is available, and abort with error
if not.