Fix Contrast in SVN Diagram
The `svn-dcommit.svg` diagram employed a too light color for the D'
element, to show it's previous state. In order to keep the colors
accessible to the visually impaired, I've colored the text in black
(like the rest) and used a dotted border on the box instead, which
should be interpreted as having the same meaning.

Fixes #17.