Update Rakefile
Add custom method for invoking Asciidoctor via CLI instead of using the
Asciidoctor API, so that we can set `failure-level` to WARN and intercept
potentially malformed documents. In order to ensure that documents which
raised a warning are viewed as outdated by Rake, set their creation and
last modified dates to Epoch 00:00:00 if Asciidoctor exits with an error.