Diagrams: Fix Striped Boxes
Remove from the striped boxed in `pull.dia` the 'M' letter from the
diagram box in the background, which was causing a "blurred letter"
effect due to non-perfect alignment with the single 'M' text-label
in the foreground.

Also remove a stray triangle (the arrow head of a zero-length line) that
was accidentally included in the striped box objects.

(see #32)