Chunk Web Edition
Render the website edition of the book as a multi-part document using
a tweaked version of @wshito's asciidoctor-chunker.

Move the single-file standalone version to a website subfolder, so end
users can download it:

    /docs/download/Git-Book.html

Update and polish the `Rakefile` to handle the new toolchain.

Update READMEs accordingly, and add download raw-link to main README.

(closes #35)